Change `mc replicate add|update` commands to integrate remote targets creation.
Deprecate `mc admin bucket remote` commands
