+++ cmd/admin-bucket-remote-main.go
@@ -22,8 +22,8 @@ import "github.com/minio/cli"
 var adminBucketRemoteSubcommands = []cli.Command{
 	adminBucketRemoteAddCmd,
 	adminBucketRemoteEditCmd,
-	adminBucketRemoteListCmd,
 	adminBucketRemoteRmCmd,
+
 	adminBwInfoCmd,
 }
