+++ cmd/replicate-add.go
@@ -20,21 +20,28 @@ package cmd
 import (
 	"context"
 	"fmt"
+	"net/url"
+	"path"
 	"strconv"
 	"strings"
+	"time"
 
+	"github.com/dustin/go-humanize"
 	"github.com/fatih/color"
 	"github.com/minio/cli"
 	json "github.com/minio/colorjson"
+	"github.com/minio/madmin-go/v2"
 	"github.com/minio/mc/pkg/probe"
 	"github.com/minio/minio-go/v7/pkg/replication"
+	"github.com/minio/minio-go/v7/pkg/s3utils"
 	"github.com/minio/pkg/console"
 )
 
 var replicateAddFlags = []cli.Flag{
 	cli.StringFlag{
-		Name:  "arn",
-		Usage: "unique role ARN",
+		Name:   "arn",
+		Usage:  "unique role ARN",
+		Hidden: true,
 	},
 	cli.StringFlag{
 		Name:  "id",
@@ -65,6 +72,32 @@ var replicateAddFlags = []cli.Flag{
 		Value: `delete-marker,delete,existing-objects,metadata-sync`,
 		Usage: `comma separated list to enable replication of soft deletes, permanent deletes, existing objects and metadata sync`,
 	},
+	cli.StringFlag{
+		Name:  "path",
+		Value: "auto",
+		Usage: "bucket path lookup supported by the server. Valid options are '[on,off,auto]'",
+	},
+	cli.StringFlag{
+		Name:  "region",
+		Usage: "region of the destination bucket (optional)",
+	},
+	cli.StringFlag{
+		Name:  "bandwidth",
+		Usage: "Set bandwidth limit in bits per second (K,B,G,T for metric and Ki,Bi,Gi,Ti for IEC units)",
+	},
+	cli.BoolFlag{
+		Name:  "sync",
+		Usage: "enable synchronous replication for this target. Default is async",
+	},
+	cli.UintFlag{
+		Name:  "healthcheck-seconds",
+		Usage: "health check interval in seconds",
+		Value: 60,
+	},
+	cli.BoolFlag{
+		Name:  "disable-proxy",
+		Usage: "disable proxying in active-active replication. If unset, default behavior is to proxy",
+	},
 }
 
 var replicateAddCmd = cli.Command{
@@ -85,23 +118,23 @@ FLAGS:
  {{end}}
 EXAMPLES:
  1. Add replication configuration rule on bucket "mybucket" for alias "myminio" to replicate all operations in an active-active replication setup.
-    {{.Prompt}} {{.HelpName}} myminio/mybucket --remote-bucket "arn:minio:replica::c5be6b16-769d-432a-9ef1-4567081f3566:destbucket" \
-        --priority 1
+    {{.Prompt}} {{.HelpName}} myminio/mybucket --remote-bucket https://foobar:foo12345@minio.siteb.example.com/targetbucket \
+        --priority 1 
 
  2. Add replication configuration rule on bucket "mybucket" for alias "myminio" to replicate all objects with tags
-    "key1=value1, key2=value2" to destbucket.
-    {{.Prompt}} {{.HelpName}} myminio/mybucket --remote-bucket "arn:minio:replica::c5be6b16-769d-432a-9ef1-4567081f3566:destbucket" \
-        --tags "key1=value1&key2=value2" \
+    "key1=value1, key2=value2" to targetbucket synchronously with bandwidth set to 2 gigabits per second. 
+    {{.Prompt}} {{.HelpName}} myminio/mybucket --remote-bucket https://foobar:foo12345@minio.siteb.example.com/targetbucket  \
+        --tags "key1=value1&key2=value2" --bandwidth "2G" --sync \
         --priority 1
 
  3. Disable a replication configuration rule on bucket "mybucket" for alias "myminio".
-    {{.Prompt}} {{.HelpName}} myminio/mybucket --remote-bucket "arn:minio:replica::c5be6b16-769d-432a-9ef1-4567081f3566:destbucket" \
+    {{.Prompt}} {{.HelpName}} myminio/mybucket --remote-bucket https://foobar:foo12345@minio.siteb.example.com/targetbucket  \
         --tags "key1=value1&key2=value2" \
         --priority 1 --disable
 
  4. Add replication configuration rule with existing object replication, delete marker replication and versioned deletes
     enabled on bucket "mybucket" for alias "myminio".
-    {{.Prompt}} {{.HelpName}} myminio/mybucket --remote-bucket "arn:minio:replica::c5be6b16-769d-432a-9ef1-4567081f3566:destbucket" \
+    {{.Prompt}} {{.HelpName}} myminio/mybucket --remote-bucket https://foobar:foo12345@minio.siteb.example.com/targetbucket  \
         --replicate "existing-objects,delete,delete-marker" \
         --priority 1
 `,
@@ -143,6 +176,91 @@ func (l replicateAddMessage) String() string {
 	return console.Colorize("replicateAddMessage", "Replication configuration rule applied to "+l.URL+" successfully.")
 }
 
+func extractCredentialURL(argURL string) (accessKey, secretKey string, u *url.URL) {
+	var parsedURL string
+	if hostKeyTokens.MatchString(argURL) {
+		fatalIf(errInvalidArgument().Trace(argURL), "temporary tokens are not allowed for remote targets")
+	}
+	if hostKeys.MatchString(argURL) {
+		parts := hostKeys.FindStringSubmatch(argURL)
+		if len(parts) != 5 {
+			fatalIf(errInvalidArgument().Trace(argURL), "Unsupported remote target format, please check --help")
+		}
+		accessKey = parts[2]
+		secretKey = parts[3]
+		parsedURL = fmt.Sprintf("%s%s", parts[1], parts[4])
+	}
+	var e error
+	if parsedURL == "" {
+		fatalIf(errInvalidArgument().Trace(argURL), "No valid credentials were detected")
+	}
+	u, e = url.Parse(parsedURL)
+	if e != nil {
+		fatalIf(errInvalidArgument().Trace(parsedURL), "Unsupported URL format %v", e)
+	}
+
+	return accessKey, secretKey, u
+}
+
+// fetchRemoteTarget - returns the dest bucket, dest endpoint, access and secret key
+func fetchRemoteTarget(cli *cli.Context) (sourceBucket string, bktTarget *madmin.BucketTarget) {
+	args := cli.Args()
+	if !cli.IsSet("remote-bucket") {
+		fatalIf(probe.NewError(fmt.Errorf("missing Remote target configuration")), "Unable to parse remote target")
+	}
+	_, sourceBucket = url2Alias(args[0])
+	p := cli.String("path")
+	if !isValidPath(p) {
+		fatalIf(errInvalidArgument().Trace(p),
+			"Unrecognized bucket path style. Valid options are `[on,off, auto]`.")
+	}
+
+	tgtURL := cli.String("remote-bucket")
+	accessKey, secretKey, u := extractCredentialURL(tgtURL)
+	var tgtBucket string
+	if u.Path != "" {
+		tgtBucket = path.Clean(u.Path[1:])
+	}
+	if e := s3utils.CheckValidBucketName(tgtBucket); e != nil {
+		fatalIf(probe.NewError(e).Trace(tgtURL), "Invalid target bucket specified")
+	}
+
+	bandwidthStr := cli.String("bandwidth")
+	bandwidth, err := getBandwidthInBytes(bandwidthStr)
+	if err != nil {
+		fatalIf(errInvalidArgument().Trace(bandwidthStr), "Invalid bandwidth number")
+	}
+	console.SetColor(cred, color.New(color.FgYellow, color.Italic))
+	creds := &madmin.Credentials{AccessKey: accessKey, SecretKey: secretKey}
+	disableproxy := cli.Bool("disable-proxy")
+	bktTarget = &madmin.BucketTarget{
+		TargetBucket:        tgtBucket,
+		Secure:              u.Scheme == "https",
+		Credentials:         creds,
+		Endpoint:            u.Host,
+		Path:                p,
+		API:                 "s3v4",
+		Type:                madmin.ServiceType("replication"),
+		Region:              cli.String("region"),
+		BandwidthLimit:      int64(bandwidth),
+		ReplicationSync:     cli.Bool("sync"),
+		DisableProxy:        disableproxy,
+		HealthCheckDuration: time.Duration(cli.Uint("healthcheck-seconds")) * time.Second,
+	}
+	return sourceBucket, bktTarget
+}
+
+func getBandwidthInBytes(bandwidthStr string) (bandwidth uint64, err error) {
+	if bandwidthStr != "" {
+		bandwidth, err = humanize.ParseBytes(bandwidthStr)
+		if err != nil {
+			return
+		}
+	}
+	bandwidth = bandwidth / 8
+	return
+}
+
 func mainReplicateAdd(cliCtx *cli.Context) error {
 	ctx, cancelReplicateAdd := context.WithCancel(globalContext)
 	defer cancelReplicateAdd()
@@ -154,6 +272,16 @@ func mainReplicateAdd(cliCtx *cli.Context) error {
 	// Get the alias parameter from cli
 	args := cliCtx.Args()
 	aliasedURL := args.Get(0)
+
+	// Create a new MinIO Admin Client
+	admclient, cerr := newAdminClient(aliasedURL)
+	fatalIf(cerr, "Unable to initialize admin connection.")
+
+	sourceBucket, bktTarget := fetchRemoteTarget(cliCtx)
+	arn, e := admclient.SetRemoteTarget(globalContext, sourceBucket, bktTarget)
+	if e != nil {
+		fatalIf(probe.NewError(e).Trace(args...), "Unable to configure remote target")
+	}
 	// Create a new Client
 	client, err := newClient(aliasedURL)
 	fatalIf(err, "Unable to initialize connection.")
@@ -185,12 +313,11 @@ func mainReplicateAdd(cliCtx *cli.Context) error {
 
 	opts := replication.Options{
 		TagString:               cliCtx.String("tags"),
-		RoleArn:                 cliCtx.String("arn"),
 		StorageClass:            cliCtx.String("storage-class"),
 		Priority:                strconv.Itoa(cliCtx.Int("priority")),
 		RuleStatus:              ruleStatus,
 		ID:                      cliCtx.String("id"),
-		DestBucket:              cliCtx.String("remote-bucket"),
+		DestBucket:              arn,
 		Op:                      replication.AddOption,
 		ReplicateDeleteMarkers:  dmReplicateStatus,
 		ReplicateDeletes:        deleteReplicationStatus,
